+++ src/Ext/Bullet/Trajectories/PhobosTrajectory.h
@@ -11,6 +11,7 @@ enum class TrajectoryFlag : int
 	Invalid = -1,
 	Straight = 0,
 	Bombard = 1,
+	FullLine = 2,
 };
 
 enum class TrajectoryCheckReturnType : int