+++ src/Ext/WeaponType/Body.cpp
@@ -48,6 +48,7 @@ void WeaponTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->FeedbackWeapon.Read(exINI, pSection, "FeedbackWeapon", true);
 	this->Laser_IsSingleColor.Read(exINI, pSection, "IsSingleColor");
 	this->Trajectory_Speed.Read(exINI, pSection, "Trajectory.Speed");
+	this->ProjectileRange.Read(exINI, pSection, "ProjectileRange");
 }
 
 template <typename T>
@@ -71,6 +72,7 @@ void WeaponTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->FeedbackWeapon)
 		.Process(this->Laser_IsSingleColor)
 		.Process(this->Trajectory_Speed)
+		.Process(this->ProjectileRange)
 		;
 };